1.21 BEGONE@!!!
There is now ammunition depleted alerts if you have a bow/crossbpw in your hand and you run out of arrows or if your flying with an elytra and run out of fireworks or you run out of snowballs to throw ):

this also includes a lot of the beeps and boops of the hev suit.

better cooldown

i think thats all?